[refactor]: move stream gen classes into lib/streams/generators
