+++ lib/streams/generators/stock_stream_generator.dart
