+++ lib/streams/generators/user_info_generator.dart
