+++ lib/streams/global_streams.dart
@@ -12,8 +12,8 @@ import 'package:dalal_street_client/proto_build/datastreams/Subscribe.pb.dart';
 import 'package:dalal_street_client/proto_build/datastreams/Transactions.pb.dart';
 import 'package:dalal_street_client/proto_build/models/Stock.pb.dart';
 import 'package:dalal_street_client/proto_build/models/User.pb.dart';
-import 'package:dalal_street_client/streams/user_info_generator.dart';
-import 'package:dalal_street_client/streams/stock_stream_generator.dart';
+import 'package:dalal_street_client/streams/generators/user_info_generator.dart';
+import 'package:dalal_street_client/streams/generators/stock_stream_generator.dart';
 import 'package:dalal_street_client/utils/convert.dart';
 import 'package:equatable/equatable.dart';
 import 'package:rxdart/streams.dart';
